Number of cards to hit lands
Runs a simulation to give the average number of draws required to hit a given
number of lands. e.g. if you have 12 swamps in a deck, and want to know on
average how many cards you need to draw (including opening hand) to see
two of them for a double-black spell, this will tell you.